Add miscellaneous minor fix.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5580 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/it_use3.cc
@@ -985,7 +985,7 @@ static bool box_of_beasts()
                                    : MONS_GIANT_LIZARD);
 
         beh_type beha = BEH_FRIENDLY;
-        unsigned int hitting = you.pet_target;
+        unsigned short hitting = you.pet_target;
 
         if (one_chance_in(you.skills[SK_EVOCATIONS] + 5))
         {
